+++ strider/track.py
@@ -1,12 +1,14 @@
 from typing import Tuple, Set, Optional
 
+from functools import total_ordering
+
 from sortedcontainers import SortedDict
 
 
 class PointOverrideException(Exception):
     pass
 
-
+@total_ordering
 class Track:
     @staticmethod
     def id(t):
@@ -116,3 +118,6 @@ class Track:
             else:
                 ret.extend([', ', k, ': ', str(v)])
         return ''.join(ret)
+
+    def __lt__(self, other):
+        return self.id < other.id